Start supporting parallelism
Decorators need to change so that the user written function
remains the same.  That is:

```
@f
def g(): ...
```

should be the same as `g`.  The reason for this is that otherwise,
`g` becomes something different -- `f(g)` -- from its definition --
a problem that makes `g` unpickleable and hence unable to be used
in multiple processes.