Save target state in task-specific directory
This prevents cluttering the target state directory with targets
from different states.  Currently, only one task is allowed to run
at a given time until completion, so this isn't an issue; however,
it will be if task parallelization is supported (as opposed to
target parallelization which is currently supported).

Problems still might appear when unexpected errors occur that
cause a series of `maid.run` commands to run.  In this case, the
previous target state will be deleted, despite their preservation
being preferrable so that the next run will continue where it
last left.